refactor: remove a layer of components from no-code
